add ClientSimulation
