+++ src/common/sim.ts
@@ -1,4 +1,4 @@
-import { Id } from './misc';
+import { Id, Input } from './misc';
 import Level from './map'; // alias to not conflict with a map collection
 import State from './state';
 import { Body, Box, Circle, Vec2, World } from 'planck-js';
@@ -116,4 +116,20 @@ function circleBody(
   });
   body.createFixture(shape);
   return body;
+}
+
+export function updatePlayerBodyFromInput(body: Body, input: Input): void {
+  // we move a player by simply increasing or decreasing its velocity in the cardinal directions
+
+  const MOVE_FORCE_STRENGTH = 10;
+  const velocity = body.getLinearVelocity();
+
+  if (input.up && !input.down) velocity.y -= MOVE_FORCE_STRENGTH;
+  else if (input.down && !input.up) velocity.y += MOVE_FORCE_STRENGTH;
+
+  if (input.left && !input.right) velocity.x -= MOVE_FORCE_STRENGTH;
+  else if (input.right && !input.left) velocity.x += MOVE_FORCE_STRENGTH;
+
+  body.setLinearVelocity(velocity);
+  body.setAwake(true);
 }
